consistent naming and fix imports
