+++ lib/extensions/componentmessage.ts
@@ -1,145 +1,125 @@
 import {
   GuildChannelOverwriteOptions,
+  MessageComponentInteraction,
   PermissionOverwriteOptions,
   EmojiIdentifierResolvable,
   WebhookEditMessageOptions,
-  CommandInteractionOption,
   DeconstructedSnowflake,
   GuildMemberResolvable,
   WebhookMessageOptions,
   AwaitMessagesOptions,
-  PermissionOverwrites,
   CreateInviteOptions,
-  CommandInteraction,
+  MessageEditOptions,
   MessageResolvable,
-  MessageAttachment,
-  CollectorFilter,
-  MessageMentions,
   MessageReaction,
-  MessagePayload,
-  UserResolvable,
+  MessageManager,
   RoleResolvable,
-  ThreadChannel,
+  UserResolvable,
+  MessagePayload,
   SnowflakeUtil,
-  Permissions,
+  ThreadChannel,
+  MessageEmbed,
   NewsChannel,
+  Permissions,
   Collection,
-  DMChannel,
   Snowflake,
+  DMChannel,
+  Webhook,
 } from "discord.js";
-import { ArgumentOptions, Command } from "@fire/lib/util/command";
-import { CommandUtil } from "@fire/lib/util/commandutil";
-import { constants } from "@fire/lib/util/constants";
-import { Language } from "@fire/lib/util/language";
 import { FireTextChannel } from "./textchannel";
 import { APIMessage } from "discord-api-types";
+import { constants } from "../util/constants";
+import { Language } from "../util/language";
 import { FireMember } from "./guildmember";
 import { FireMessage } from "./message";
-import { Fire } from "@fire/lib/Fire";
 import { FireGuild } from "./guild";
 import { FireUser } from "./user";
+import { Fire } from "../Fire";
 
 const { emojis, reactions } = constants;
+export type EphemeralMessage = { id: Snowflake; flags: number };
 
-export class SlashCommandMessage {
+export class ComponentMessage {
   realChannel?: FireTextChannel | NewsChannel | DMChannel;
-  attachments: Collection<string, MessageAttachment>;
   private snowflake: DeconstructedSnowflake;
-  slashCommand: CommandInteraction;
+  message: FireMessage | EphemeralMessage;
+  interaction: MessageComponentInteraction;
   sent: false | "ack" | "message";
   sourceMessage: FireMessage;
-  mentions: MessageMentions;
   latestResponse: Snowflake;
   private _flags: number;
   channel: FakeChannel;
-  member?: FireMember;
+  ephemeral: boolean;
+  member: FireMember;
   language: Language;
-  guild?: FireGuild;
-  util: CommandUtil;
-  command: Command;
+  customID: string;
+  guild: FireGuild;
   author: FireUser;
-  webhookID = null;
-  content: string;
   id: Snowflake;
   client: Fire;
 
-  constructor(client: Fire, command: CommandInteraction) {
+  constructor(client: Fire, button: MessageComponentInteraction) {
     this.client = client;
-    this.id = command.id;
+    this.id = button.id;
     this.snowflake = SnowflakeUtil.deconstruct(this.id);
-    this.slashCommand = command;
+    this.customID = button.customID;
+    this.interaction = button;
+    this.sent = false;
+    this.guild = button.guild as FireGuild;
+    this.realChannel = client.channels.cache.get(button.channelID) as
+      | FireTextChannel
+      | NewsChannel
+      | DMChannel;
+    this.ephemeral = button.message.flags
+      ? (button.message.flags.valueOf() & 64) != 0
+      : false;
+    this.message = this.ephemeral
+      ? (button.message as EphemeralMessage)
+      : button.message instanceof FireMessage
+      ? button.message
+      : new FireMessage(client, button.message, this.realChannel);
     if (
-      command.options?.size &&
-      command.options.first()?.type == "SUB_COMMAND"
-    ) {
-      command.commandName = `${command.commandName}-${
-        command.options.first().name
-      }`;
-      command.options = command.options.first().options;
-    }
-    this.guild = client.guilds.cache.get(command.guildID) as FireGuild;
-    this.command = this.client.getCommand(command.commandName);
-    this._flags = 0;
-    if (
-      this.guild?.tags?.slashCommands[command.commandID] == command.commandName
-    ) {
-      this.command = this.client.getCommand("tag");
-      command.options = new Collection<string, CommandInteractionOption>().set(
-        "tag",
-        {
-          name: "tag",
-          value: command.commandName,
-          type: "STRING",
-        }
-      );
-      if (this.guild.tags.ephemeral) this.flags = 64;
-    }
-    if (this.command?.ephemeral) this.flags = 64;
-    // @ts-ignore
-    this.mentions = new MessageMentions(this, [], [], false);
-    this.attachments = new Collection();
-    // @mason pls just always include user ty
-    const user = command.user ?? command.member?.user;
-    this.author =
-      (client.users.cache.get(user.id) as FireUser) ||
-      new FireUser(client, user);
-    if (!client.users.cache.has(this.author.id))
-      client.users.add(command.member ? command.member.user : command.user);
-    if (this.guild) {
+      !this.message ||
+      (!this.ephemeral &&
+        !(this.message as FireMessage).components?.find(
+          (component) =>
+            component.type == "ACTION_ROW" &&
+            component.components.find(
+              (component) => component.customID == this.customID
+            )
+        ))
+    )
+      throw new Error("Component checks failed, potential mitm/selfbot?");
+    if (button.member)
       this.member =
-        (this.guild.members.cache.get(this.author.id) as FireMember) ||
-        new FireMember(client, command.member, this.guild);
-      if (!this.guild.members.cache.has(this.member.id))
-        this.guild.members.add(command.member);
-    }
+        (this.guild.members.cache.get(button.member.user.id) as FireMember) ||
+        new FireMember(client, button.member, this.guild);
+    this.author = button.user
+      ? (client.users.cache.get(button.user.id) as FireUser) ||
+        new FireUser(client, button.user)
+      : button.member &&
+        ((client.users.cache.get(button.member.user.id) as FireUser) ||
+          new FireUser(client, button.member.user));
     this.language = this.author?.settings.has("utils.language")
       ? this.author.language.id == "en-US" && this.guild?.language.id != "en-US"
         ? this.guild?.language
         : this.author.language
       : this.guild?.language || client.getLanguage("en-US");
-    this.realChannel = this.client.channels.cache.get(
-      this.slashCommand.channelID
-    ) as FireTextChannel | NewsChannel | DMChannel;
-    this.latestResponse = "@original" as Snowflake;
-    this.sent = false;
-    this.util = new CommandUtil(this.client.commandHandler, this);
     if (!this.guild) {
-      // This will happen if a guild authorizes w/applications.commands only
-      // or if a slash command is invoked in DMs (discord/discord-api-docs #2568)
       this.channel = new FakeChannel(
         this,
         client,
-        command.id,
-        command.token,
-        command.guildID ? null : this.author.dmChannel
+        button.id,
+        button.token,
+        button.guildID ? null : this.author.dmChannel
       );
       return this;
     }
     this.channel = new FakeChannel(
       this,
       client,
-      command.id,
-      command.token,
+      button.id,
+      button.token,
       this.realChannel
     );
   }
@@ -178,52 +158,6 @@ export class SlashCommandMessage {
     return this.snowflake.timestamp;
   }
 
-  get url() {
-    if (this.sourceMessage)
-      return `https://discord.com/channels/${
-        this.guild ? this.guild.id : "@me"
-      }/${this.realChannel?.id || "0"}/${this.sourceMessage.id}`;
-    else
-      return `https://discord.com/channels/${
-        this.guild ? this.guild.id : "@me"
-      }/${this.realChannel?.id || "0"}/${this.id}`;
-  }
-
-  async generateContent() {
-    let prefix = (this.client.commandHandler.prefix as (
-      message: any
-    ) => string | string[] | Promise<string | string[]>)(this);
-    if (this.client.util.isPromise(prefix)) prefix = await prefix;
-    if (prefix instanceof Array) prefix = prefix[0].trim();
-    let content = (prefix as string) + " ";
-    content += this.command.id + " ";
-    if (this.command.args?.length && this.slashCommand.options?.size) {
-      const commandArgs = this.command.args as ArgumentOptions[];
-      const argNames = this.slashCommand.options.map((opt) => opt.name);
-      const sortedArgs = this.slashCommand.options.sort(
-        (a, b) =>
-          argNames.indexOf(a.name.toLowerCase()) -
-          argNames.indexOf(b.name.toLowerCase())
-      );
-      let args = sortedArgs.map((opt) => {
-        if (
-          commandArgs.find(
-            (arg) => arg.id == opt.name && arg.flag && arg.match == "flag"
-          ) &&
-          opt.value
-        ) {
-          const arg = commandArgs.find((arg) => arg.id == opt.name);
-          return arg.flag;
-        } else if (commandArgs.find((arg) => arg.id == opt.name && arg.flag))
-          return `--${opt.name} ${opt.value}`;
-        return opt.value;
-      });
-      content += args.join(" ");
-    }
-    this.content = content;
-    return this.content;
-  }
-
   send(key: string = "", ...args: any[]) {
     return this.channel.send(
       { content: this.language.get(key, ...args) },
@@ -234,7 +168,7 @@ export class SlashCommandMessage {
   success(
     key: string = "",
     ...args: any[]
-  ): Promise<SlashCommandMessage | MessageReaction | void> {
+  ): Promise<ComponentMessage | MessageReaction | void> {
     if (!key) {
       if (this.sourceMessage instanceof FireMessage)
         return this.sourceMessage.react(reactions.success).catch(() => {});
@@ -256,7 +190,7 @@ export class SlashCommandMessage {
   error(
     key: string = "",
     ...args: any[]
-  ): Promise<SlashCommandMessage | MessageReaction | void> {
+  ): Promise<ComponentMessage | MessageReaction | void> {
     if (!key) {
       if (this.sourceMessage instanceof FireMessage)
         return this.sourceMessage.react(reactions.error).catch(() => {});
@@ -270,18 +204,22 @@ export class SlashCommandMessage {
         });
     }
     return this.channel.send(
-      `${emojis.slashError} ${this.language.get(key, ...args)}`,
+      `${emojis.error} ${this.language.get(key, ...args)}`,
       typeof this.flags == "number" ? this.flags : 64
     );
   }
 
   async getRealMessage() {
-    if (!this.realChannel) return;
+    if (!this.realChannel || this.ephemeral) return;
     if (this.sourceMessage instanceof FireMessage) return this.sourceMessage;
 
-    let messageId = this.latestResponse as Snowflake;
+    let messageId = this.latestResponse;
     if (messageId == "@original") {
-      const message = await this.slashCommand.fetchReply().catch(() => {});
+      const message = await this.client.req
+        .webhooks(this.client.user.id, this.interaction.token)
+        .messages(messageId)
+        .get<APIMessage>()
+        .catch(() => {});
       if (message) messageId = message.id;
     }
 
@@ -292,37 +230,56 @@ export class SlashCommandMessage {
     return message;
   }
 
-  async edit(options?: WebhookEditMessageOptions | MessagePayload) {
+  async edit(
+    content: string | MessageEditOptions | MessageEmbed | MessagePayload,
+    options?: WebhookEditMessageOptions & { embed?: MessageEmbed }
+  ) {
     let apiMessage: MessagePayload;
 
-    if (options instanceof MessagePayload) apiMessage = options.resolveData();
+    if (content instanceof MessageEmbed) {
+      options = {
+        ...options,
+        embeds: [content],
+      };
+      content = null;
+    }
+
+    if (options?.embed) {
+      options.embeds = [options.embed];
+      delete options.embed;
+    }
+
+    if (content instanceof MessagePayload) apiMessage = content.resolveData();
     else {
       apiMessage = MessagePayload.create(
-        this.slashCommand,
+        new Webhook(this.client, null), // needed to make isWebhook true for embeds array
+        content as string,
         options
       ).resolveData();
     }
 
-    const { data, files } = (await apiMessage.resolveFiles()) as {
+    const { data } = (await apiMessage.resolveFiles()) as {
       data: any;
       files: any[];
     };
 
+    data.flags = this.flags;
+
     await this.client.req
-      .webhooks(this.client.user.id, this.slashCommand.token)
-      .messages(this.latestResponse)
+      .webhooks(this.client.user.id, this.interaction.token)
+      .messages(this.latestResponse ?? "@original")
       .patch({
         data,
-        files,
       })
       .catch(() => {});
     return this;
   }
 
-  async delete() {
+  async delete(id?: string) {
+    if (this.ephemeral) return;
     await this.client.req
-      .webhooks(this.client.user.id, this.slashCommand.token)
-      .messages(this.latestResponse)
+      .webhooks(this.client.user.id, this.interaction.token)
+      .messages(id ?? this.latestResponse ?? "@original")
       .delete()
       .catch(() => {});
   }
@@ -334,31 +291,20 @@ export class SlashCommandMessage {
 
     return await this.sourceMessage.react(emoji);
   }
-
-  hasExperiment(id: number, bucket: number) {
-    // if (this.client.config.dev) return true;
-    const experiment = this.client.experiments.get(id);
-    if (!experiment) return false;
-    else if (!experiment.active) return true;
-    else if (experiment.kind == "guild" && !this.guild) return false;
-    else if (experiment.kind == "guild")
-      return this.guild.hasExperiment(id, bucket);
-    else return this.author.hasExperiment(id, bucket);
-  }
 }
 
 export class FakeChannel {
-  real: FireTextChannel | NewsChannel | ThreadChannel | DMChannel;
-  message: SlashCommandMessage;
-  guild?: FireGuild;
+  real: FireTextChannel | ThreadChannel | NewsChannel | DMChannel;
+  messages: MessageManager;
+  message: ComponentMessage;
   token: string;
-  id: Snowflake;
   client: Fire;
+  id: string;
 
   constructor(
-    message: SlashCommandMessage,
+    message: ComponentMessage,
     client: Fire,
-    id: Snowflake,
+    id: string,
     token: string,
     real?: FireTextChannel | NewsChannel | DMChannel
   ) {
@@ -367,24 +313,11 @@ export class FakeChannel {
     this.token = token;
     this.client = client;
     this.message = message;
-
-    if (!(real instanceof DMChannel) && real?.guild)
-      this.guild = real.guild as FireGuild;
-    else if (this.message.guild) this.guild = this.message.guild;
+    this.messages = real?.messages;
   }
 
   get flags() {
     return this.message.flags;
-  }
-
-  get permissionOverwrites() {
-    return this.real instanceof DMChannel || this.real instanceof ThreadChannel
-      ? new Collection<string, PermissionOverwrites>()
-      : this.real.permissionOverwrites;
-  }
-
-  get messages() {
-    return this.real.messages;
   }
 
   toString() {
@@ -441,13 +374,14 @@ export class FakeChannel {
   }
 
   // Acknowledges without sending a message
-  async ack(ephemeral = false) {
-    if (ephemeral || (this.flags & 64) != 0) return;
-    await this.message.slashCommand
-      .defer({ ephemeral: !!((this.flags & 64) == 64) })
+  async ack() {
+    await this.client.req
+      .interactions(this.id)(this.token)
+      .callback.post({
+        data: { type: 6 },
+      })
       .then(() => {
         this.message.sent = "ack";
-        this.message.getRealMessage().catch(() => {});
       })
       .catch(() => (this.message.sent = "ack"));
   }
@@ -458,13 +392,13 @@ export class FakeChannel {
       | MessagePayload
       | (WebhookMessageOptions & { split?: false }),
     flags?: number // Used for success/error, can also be set
-  ): Promise<SlashCommandMessage> {
+  ): Promise<ComponentMessage> {
     let apiMessage: MessagePayload;
 
     if (options instanceof MessagePayload) apiMessage = options.resolveData();
     else {
       apiMessage = MessagePayload.create(
-        this.message.slashCommand,
+        new Webhook(this.client, null), // needed to make isWebhook true for embeds array
         options
       ).resolveData();
     }
@@ -477,7 +411,6 @@ export class FakeChannel {
     data.flags = this.flags;
     if (typeof flags == "number") data.flags = flags;
 
-    // embeds in ephemeral wen eta
     if (
       (files?.length || this.real instanceof DMChannel) &&
       (data.flags & 64) == 64
@@ -496,21 +429,10 @@ export class FakeChannel {
         })
         .then(() => {
           this.message.sent = "message";
+          this.message.latestResponse = "@original" as Snowflake;
         })
         .catch(() => {});
-    else if (this.message.sent == "ack") {
-      await this.client.req
-        .webhooks(this.client.user.id)(this.token)
-        .messages("@original")
-        .patch({
-          data,
-          files,
-        })
-        .then(() => {
-          this.message.sent = "message";
-        })
-        .catch(() => {});
-    } else {
+    else {
       const message = await this.client.req
         .webhooks(this.client.user.id)(this.token)
         .post<APIMessage>({
@@ -518,13 +440,74 @@ export class FakeChannel {
           files,
           query: { wait: true },
         })
-        .then((message) => {
-          this.message.sent = "message";
-          return message;
-        })
         .catch(() => {});
-      if (message && message.id) this.message.latestResponse = message.id;
+      if (message && message.id && this.message.latestResponse == "@original")
+        this.message.latestResponse = message.id;
+      else this.message.latestResponse = "@original" as Snowflake;
     }
+    this.message.getRealMessage().catch(() => {});
+    return this.message;
+  }
+
+  async update(
+    content: string | MessagePayload | MessageEmbed,
+    options?: WebhookMessageOptions & { embed?: MessageEmbed },
+    flags?: number // Used for success/error, can also be set
+  ): Promise<ComponentMessage> {
+    if (this.message.sent) return; // can only update with initial response
+
+    let apiMessage: MessagePayload;
+
+    if (content instanceof MessageEmbed) {
+      options = {
+        ...options,
+        embeds: [content],
+      };
+      content = null;
+    }
+
+    if (options?.embed) {
+      options.embeds = [options.embed];
+      delete options.embed;
+    }
+
+    if (content instanceof MessagePayload) apiMessage = content.resolveData();
+    else {
+      apiMessage = MessagePayload.create(
+        new Webhook(this.client, null), // needed to make isWebhook true for embeds array
+        content as string,
+        options
+      ).resolveData();
+    }
+
+    const { data, files } = (await apiMessage.resolveFiles()) as {
+      data: any;
+      files: any[];
+    };
+
+    data.flags = this.flags;
+    if (typeof flags == "number") data.flags = flags;
+
+    if (
+      (files?.length || this.real instanceof DMChannel) &&
+      (data.flags & 64) == 64
+    )
+      data.flags -= 64;
+
+    await this.client.req
+      .interactions(this.id)(this.token)
+      .callback.post({
+        data: {
+          type: 7,
+          data,
+        },
+        files,
+      })
+      .then(() => {
+        this.message.sent = "message";
+        this.message.latestResponse = "@original" as Snowflake;
+      })
+      .catch(() => {});
     this.message.getRealMessage().catch(() => {});
     return this.message;
   }
