+++ src/commands/Main/ping.ts
@@ -1,6 +1,6 @@
-import { SlashCommandMessage } from "@fire/lib/extensions/slashCommandMessage";
-import { DiscordAPIError, MessageEmbed, Permissions } from "discord.js";
+import { SlashCommandMessage } from "@fire/lib/extensions/slashcommandmessage";
 import { FireMessage } from "@fire/lib/extensions/message";
+import { MessageEmbed, Permissions } from "discord.js";
 import { Language } from "@fire/lib/util/language";
 import { Command } from "@fire/lib/util/command";
