+++ src/listeners/INTERACTION_CREATE.ts
@@ -1,12 +1,10 @@
 // The case of the file name is just to signify that
 // this is listening to an event directly from the gateway
 
-import { Button, Interaction } from "@fire/lib/interfaces/interactions";
-import { ButtonMessage } from "@fire/lib/extensions/buttonMessage";
+import { Interaction } from "@fire/lib/interfaces/interactions";
 import { FireGuild } from "@fire/lib/extensions/guild";
 import { constants } from "@fire/lib/util/constants";
 import { Listener } from "@fire/lib/util/listener";
-import { Scope } from "@sentry/node";
 
 const { emojis } = constants;
 
@@ -21,7 +19,7 @@ export default class InteractionCreate extends Listener {
   async exec(interaction: Interaction) {
     if (!interaction) return;
     if (this.blacklistCheck(interaction)) return;
-    // slash command, use client interaction event
+    // slash command or message component, use client interaction event
     else if (interaction.type == 2 || interaction.type == 3) return;
     else {
       const haste = await this.client.util.haste(
